+++ src/pattern/pattern_object.rs
@@ -1,0 +1,9 @@
+
+
+/// on what the search applies
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum PatternObject {
+    FileName,
+    FileSubpath,
+    FileContent,
+}