+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### next version
 [documentation of this version](https://dystroy.org/broot-master/)
 #### Major feature: new input syntax - Breaking Change
-- new search modes: fuzzy or regex on sub-paths (the path starting from the displayed root)
+- new search modes: fuzzy or regex on sub-paths (the path starting from the displayed root), search in file content
 - it's possible to configure how search modes are selected in config
 #### Minor changes:
 - tab goes to next direct match when there's no verb in input - Fix #234